Add backup flags to AuthenticatorData
In the working draft of the specification, bits 3 and 4 of the flags
field of the Authenticator Data structure now correspond to the backup
eligibility and backup state of the credential respectively.

https://w3c.github.io/webauthn/#sctn-authenticator-data

These flags are exposed through the flag_backup_eligible and
flag_credential_backed_up fields of the AuthenticatorData struct.